requirements.txt
# Temel Kütüphaneler
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.3

# Veri İşleme & ML
scikit-learn==1.3.0
statsmodels==0.14.0
xgboost==1.7.6
optuna==3.3.0
shap==0.42.1
river==0.15.1

# NLP & Duygu Analizi
transformers==4.33.3
torch==2.0.1

# Arayüz & Görselleştirme
matplotlib==3.7.2
seaborn==0.12.2
tk==0.1.0

# Zaman Serisi & Finans
ta==0.10.2
python-binance==1.0.19
alpha_vantage==2.3.1

# Diğer
tqdm==4.65.0
loguru==0.7.0